hard reset
Started over. Trying to get the general layout functioning. each project is a row that contains a headline, subhead (might remove), project description, and project number (still figuring out how to get the project number on the left side of the title). Each project will have a background image that displays in the background when you hover over the title.