overlay and skew implemented
So the overlay and skew on scroll have been implemented but still need refining. I'm still figuring out how to load images into a slide show based on the project that was clicked. The overlay is also supposed to have a horizontal scroll, but I need to implement horizontal scroll wheel functionality as well as make the skew.js work on horizontal scroll